FLAG: unsure what these changes are.  I must have gone a while without commiting.  Either way this works in current version for the experiment omr2stim for the vglut2:rschrmine experiment and python 3.12

+++ bruker_images.py
@@ -140,7 +140,7 @@ def bruker_img_organization(folder_path, testkey = 'Cycle', safe=False, single_p
             for tiffany in fls:
                 with tifffile.TiffFile(tiffany) as tiff:
                     for page in tiff.pages:
-                        output.write(page.asarray().astype("uint16"),
+                        output.write(page.asarray().astype("uint32"),
                                     contiguous=True)
